+++ contracts/interfaces/IStaking.sol
@@ -6,8 +6,8 @@ import "@openzeppelin/contracts/access/Ownable.sol";
 interface IStaking {
 
     function getEpochId(uint timestamp) external view returns (uint); // get epoch id
-    function getEpochUserBalance(address user, address token, uint epoch) external view returns(uint);
-    function getEpochPoolSize(address token, uint epoch) external view returns (uint);
+    function getEpochUserBalance(address user, address token, uint128 epoch) external view returns(uint);
+    function getEpochPoolSize(address token, uint128 epoch) external view returns (uint);
     function epoch1Start() external view returns (uint);
     function epochDuration() external view returns (uint);
 }
